Add option to change MDL theme.
This commit adds the ability to set the MDL color theme using the
mdlPrimary and mdlSecondary parameters in config.toml. It also switches to
using the official MDL CDN link for CSS and JS and upgrades MDL from 1.0.5
to 1.1.3.

Unfortunately, this change requires the removal of CSS for highlighting
navigation links during hover. Maybe one day MDL will provide a
mdl-hover__accent CSS class to solve this problem.